Extract weekly event aggregation into stable fair-events
The digest planned in #916 needs the 7-day event aggregation that today
only exists inside fair-events-experimental's WeeklyEventsController.
Moving it into FairEvents\Services\WeeklyEventsProvider lets stable
fair-audience reuse it without depending on the experimental plugin.
The experimental controller becomes a thin wrapper with no REST
behavior change.

Refs #916
